Support optional scroller in scrollIntoView
There currently is a typescript error when doing something like this:

```js
scrollIntoView(myElement, {behavior: 'smooth'})
```

Because the second argument is expected to be the element that you want
to do the scrolling. However, the API does allow for the scroller (2nd arg)
to be optional. So this just changes so that TS respects the above usage.