Support optional scroller in scrollIntoView (#40)
Currently TypeScript produces an error when doing something like this:

```js
scrollIntoView(myElement, {behavior: 'smooth'})
```

The error is:

```bash
Argument of type '{ behavior: string; }' is not assignable to parameter of type 'Element'.
  Object literal may only specify known properties, and 'behavior' does not exist in type 'Element'.ts(2345)
```

Because the second argument is expected to be the element that you want to do the scrolling. However, the API does allow for the scroller (2nd arg) to be optional. So this just changes so that TS respects the above usage.

(other changes are just prettier formatting goin' ham) :grin: